+++ crates/pgt_completions/src/providers/columns.rs
@@ -1,21 +1,17 @@
 use crate::{
-    CompletionItemKind,
-    builder::{CompletionBuilder, PossibleCompletionItem},
-    context::CompletionContext,
-    relevance::{CompletionRelevanceData, filtering::CompletionFilter, scoring::CompletionScore},
+    CompletionItem, CompletionItemKind, builder::CompletionBuilder, context::CompletionContext,
+    relevance::CompletionRelevanceData,
 };
 
-pub fn complete_columns<'a>(ctx: &CompletionContext<'a>, builder: &mut CompletionBuilder<'a>) {
+pub fn complete_columns(ctx: &CompletionContext, builder: &mut CompletionBuilder) {
     let available_columns = &ctx.schema_cache.columns;
 
     for col in available_columns {
-        let relevance = CompletionRelevanceData::Column(col);
-
-        let item = PossibleCompletionItem {
+        let item = CompletionItem {
             label: col.name.clone(),
-            score: CompletionScore::from(relevance.clone()),
-            filter: CompletionFilter::from(relevance),
+            score: CompletionRelevanceData::Column(col).get_score(ctx),
             description: format!("Table: {}.{}", col.schema_name, col.table_name),
+            preselected: false,
             kind: CompletionItemKind::Column,
         };
 
@@ -26,10 +22,8 @@ pub fn complete_columns<'a>(ctx: &CompletionContext<'a>, builder: &mut Completio
 #[cfg(test)]
 mod tests {
     use crate::{
-        CompletionItem, CompletionItemKind, complete,
-        test_helper::{
-            CURSOR_POS, InputQuery, get_test_deps, get_test_params, test_against_connection_string,
-        },
+        CompletionItem, complete,
+        test_helper::{CURSOR_POS, InputQuery, get_test_deps, get_test_params},
     };
 
     struct TestCase {
@@ -210,8 +204,7 @@ mod tests {
         let has_column_in_first_four = |col: &'static str| {
             first_four
                 .iter()
-                .find(|compl_item| compl_item.label.as_str() == col)
-                .is_some()
+                .any(|compl_item| compl_item.label.as_str() == col)
         };
 
         assert!(
@@ -230,101 +223,5 @@ mod tests {
             has_column_in_first_four("email"),
             "`email` not present in first four completion items."
         );
-    }
-
-    #[tokio::test]
-    async fn ignores_cols_in_from_clause() {
-        let setup = r#"
-        create schema private;
-
-        create table private.users (
-            id serial primary key,
-            name text,
-            address text,
-            email text
-        );
-    "#;
-
-        let test_case = TestCase {
-            message: "suggests user created tables first",
-            query: format!(r#"select * from private.{}"#, CURSOR_POS),
-            label: "",
-            description: "",
-        };
-
-        let (tree, cache) = get_test_deps(setup, test_case.get_input_query()).await;
-        let params = get_test_params(&tree, &cache, test_case.get_input_query());
-        let results = complete(params);
-
-        assert!(
-            !results
-                .into_iter()
-                .any(|item| item.kind == CompletionItemKind::Column)
-        );
-    }
-
-    #[tokio::test]
-    async fn prefers_columns_of_mentioned_tables() {
-        let setup = r#"
-        create schema private;
-
-        create table private.users (
-            id1 serial primary key,
-            name1 text,
-            address1 text,
-            email1 text
-        );
-
-        create table public.users (
-            id2 serial primary key,
-            name2 text,
-            address2 text,
-            email2 text
-        );
-    "#;
-
-        {
-            let test_case = TestCase {
-                message: "",
-                query: format!(r#"select {} from users"#, CURSOR_POS),
-                label: "suggests from table",
-                description: "",
-            };
-
-            let (tree, cache) = get_test_deps(setup, test_case.get_input_query()).await;
-            let params = get_test_params(&tree, &cache, test_case.get_input_query());
-            let results = complete(params);
-
-            assert_eq!(
-                results
-                    .into_iter()
-                    .take(4)
-                    .map(|item| item.label)
-                    .collect::<Vec<String>>(),
-                vec!["address2", "email2", "id2", "name2"]
-            );
-        }
-
-        {
-            let test_case = TestCase {
-                message: "",
-                query: format!(r#"select {} from private.users"#, CURSOR_POS),
-                label: "suggests from table",
-                description: "",
-            };
-
-            let (tree, cache) = get_test_deps(setup, test_case.get_input_query()).await;
-            let params = get_test_params(&tree, &cache, test_case.get_input_query());
-            let results = complete(params);
-
-            assert_eq!(
-                results
-                    .into_iter()
-                    .take(4)
-                    .map(|item| item.label)
-                    .collect::<Vec<String>>(),
-                vec!["address1", "email1", "id1", "name1"]
-            );
-        }
     }
 }
